[CBRD-24807] Refactor type evaluation of functions (#4382)
http://jira.cubrid.org/browse/CBRD-24807

- Rename FUNC_TYPE -> FUNC_CODE
- Move FUNC_CODE from storage_common.h to a new file
- Rename pt_eval_function_type_old () -> pt_eval_function_type_aggregate (): only aggregate functions are not migrated to the new routine yet (func_type::Node class)
- Trivial changes
