+++ src/parser/func_type.hpp
@@ -114,7 +114,13 @@ namespace func_type
   }; //class Node
 } //namespace func_type
 
+PT_NODE *pt_eval_function_type_aggregate (PARSER_CONTEXT *parser, PT_NODE *node);
+
 bool pt_are_equivalent_types (const PT_ARG_TYPE def_type, const PT_TYPE_ENUM op_type);
 PT_TYPE_ENUM pt_get_equivalent_type (const PT_ARG_TYPE def_type, const PT_TYPE_ENUM arg_type);
+
+bool pt_is_function_unsupported (FUNC_CODE fcode);
+bool pt_is_function_no_arg (FUNC_CODE fcode);
+bool pt_is_function_new_type_checking (FUNC_CODE fcode);
 
 #endif
