+++ src/parser/parse_tree.h
@@ -36,6 +36,7 @@
 #include "compile_context.h"
 #include "config.h"
 #include "cursor.h"
+#include "db_function.hpp"
 #include "json_table_def.h"
 #include "message_catalog.h"
 #include "string_opfunc.h"
@@ -2426,7 +2427,7 @@ struct pt_file_path_info
 struct pt_function_info
 {
   PT_NODE *arg_list;		/* PT_EXPR(list) */
-  FUNC_TYPE function_type;	/* PT_COUNT, PT_AVG, ... */
+  FUNC_CODE function_type;	/* PT_COUNT, PT_AVG, ... */
   PT_MISC_TYPE all_or_distinct;	/* will be PT_ALL or PT_DISTINCT */
   const char *generic_name;	/* only for type PT_GENERIC */
   char hidden_column;		/* used for updates and deletes for the class OID column */
